deprecating TextInput and creating InputText
