+++ apps/docs/content/3-components/2-library/input-password.mdx
@@ -43,10 +43,10 @@ The `InputPassword` component provides a secure password input field with an int
     <fieldset class="">
     <label class="gi-text-md gi-label gi-mb-1" for="password-input-id">Password</label>
         <div class="gi-hint-text-md gi-hint-text gi-mb-1">Hint: Enter your password securely.</div>
-        <div class="gi-text-input-container">
-            <div class="gi-text-input-inner">
-                <input data-end-element="true" class="gi-text-input" data-testid="password-input-id" id="password-input-id" type="password">
-                <div class="gi-text-input-end-element" data-suffix="false"><button role="button" data-testid="govieIconButton-dark-flat-small-undefined" class="gi-btn gi-btn-flat-dark gi-icon-btn-small undefined "><span data-testid="govie-icon" role="presentation" class="gi-block material-symbols-outlined" style="font-size: 16px;">visibility</span></button></div>
+        <div class="gi-input-text-container">
+            <div class="gi-input-text-inner">
+                <input data-end-element="true" class="gi-input-text" data-testid="password-input-id" id="password-input-id" type="password">
+                <div class="gi-input-text-end-element" data-suffix="false"><button role="button" data-testid="govieIconButton-dark-flat-small-undefined" class="gi-btn gi-btn-flat-dark gi-icon-btn-small undefined "><span data-testid="govie-icon" role="presentation" class="gi-block material-symbols-outlined" style="font-size: 16px;">visibility</span></button></div>
             </div>
         </div>
     </fieldset>
