+++ apps/docs/content/4-patterns/1-ask-users-for/1-addresses.mdx
@@ -31,7 +31,7 @@ This guidance is for government teams creating online address forms for digital 
                 <div class="gi-hint-text-md gi-hint-text gi-mb-1"
                     aria-label="Street address, P.O. box, company name, c/o">Street address, P.O. box, company name, c/o
                 </div>
-                <div class="gi-text-input-container"><input class="gi-text-input" id="address-line-1"></div>
+                <div class="gi-input-text-container"><input class="gi-input-text" id="address-line-1"></div>
             </fieldset>
         </div>
         <div class="gi-flex gi-w-full gi-justify-start gi-items-start gi-flex-col md:gi-flex-row gi-gap-3 gi-flex-nowrap"
@@ -41,7 +41,7 @@ This guidance is for government teams creating online address forms for digital 
                 <div class="gi-hint-text-md gi-hint-text gi-mb-1"
                     aria-label="Apartment, suite, unit, building, floor, etc. (optional)">Apartment, suite, unit,
                     building, floor, etc. (optional)</div>
-                <div class="gi-text-input-container"><input class="gi-text-input" id="address-line-2"></div>
+                <div class="gi-input-text-container"><input class="gi-input-text" id="address-line-2"></div>
             </fieldset>
         </div>
         <div class="gi-flex gi-w-full gi-justify-start gi-items-start gi-flex-col md:gi-flex-row gi-gap-3 gi-flex-nowrap"
@@ -49,12 +49,12 @@ This guidance is for government teams creating online address forms for digital 
             <fieldset class="" data-testid="govie-stack-item-0"><label class="gi-text-md gi-label gi-mb-1"
                     aria-label="Town or City">Town or City</label>
                 <div class="gi-hint-text-md gi-hint-text gi-mb-1" aria-label="Your town or city">Your town or city</div>
-                <div class="gi-text-input-container"><input class="gi-text-input" id="town-city"></div>
+                <div class="gi-input-text-container"><input class="gi-input-text" id="town-city"></div>
             </fieldset>
             <fieldset class="" data-testid="govie-stack-item-1"><label class="gi-text-md gi-label gi-mb-1"
                     aria-label="Country">Country</label>
                 <div class="gi-hint-text-md gi-hint-text gi-mb-1" aria-label="Your country">Your country</div>
-                <div class="gi-text-input-container"><input class="gi-text-input" id="country"></div>
+                <div class="gi-input-text-container"><input class="gi-input-text" id="country"></div>
             </fieldset>
         </div>
         <div class="gi-w-1/2" data-testid="govie-stack-item-3">
@@ -64,7 +64,7 @@ This guidance is for government teams creating online address forms for digital 
                         aria-label="Postcode">Postcode</label>
                     <div class="gi-hint-text-md gi-hint-text gi-mb-1" aria-label="Your postcode or ZIP code">Your
                         postcode or ZIP code</div>
-                    <div class="gi-text-input-container"><input class="gi-text-input" id="postcode"
+                    <div class="gi-input-text-container"><input class="gi-input-text" id="postcode"
                             maxlength="10"></div>
                 </fieldset>
             </div>
@@ -124,6 +124,6 @@ This guidance is for government teams creating online address forms for digital 
         </Stack>
       </div>
     ```
+
   </TabPanel>
 </Tabs>
-
