+++ apps/docs/content/4-patterns/2-pages/1-basic-page.mdx
@@ -88,7 +88,6 @@ A template for a basic page
                     search
                   </span>
                   <span
-                
                     role="presentation"
                     style="
                       font-variation-settings:
@@ -138,12 +137,12 @@ A template for a basic page
       <form action="/search_page" class="gi-max-w-md gi-mx-auto">
         <h4 class="gi-heading-sm">Search the website</h4>
         <div class="gi-flex gi-items-end gi-mt-4">
-          <div class="gi-text-input-container  gi-flex-auto">
-            <div class="gi-text-input-container"><input placeholder="Enter search term" id="search" type="text"
-                class="gi-border-gray-950 gi-w-full gi-text-input" name="search_query"
+          <div class="gi-input-text-container  gi-flex-auto">
+            <div class="gi-input-text-container"><input placeholder="Enter search term" id="search" type="text"
+                class="gi-border-gray-950 gi-w-full gi-input-text" name="search_query"
                 aria-label="Search the website"></div>
           </div>
-          <button 
+          <button
             class="gi-btn gi-btn-primary gi-btn-regular gi-ml-1 gi-flex-none">Search<spa
               aria-label="Search" role="img" style="
               font-variation-settings:
@@ -161,7 +160,7 @@ A template for a basic page
         <div class="gi-flex gi-items-center">
           <label for="MobileMenuTrigger" class="gi-header-menu-mobile-trigger">
             <span>Close</span>
-            <span role="presentation" 
+            <span role="presentation"
             style="
             font-variation-settings:
               'FILL' 0,
@@ -187,14 +186,14 @@ A template for a basic page
           <form action="/search_page" class="gi-max-w-md gi-mx-auto">
             <h4 class="gi-heading-sm">Search the website</h4>
             <div class="gi-flex gi-items-end gi-mt-4">
-              <div class="gi-text-input-container  gi-flex-auto">
-                <div class="gi-text-input-container">
-                  <input placeholder="Enter search term" id="search" type="text" 
-                    class="gi-border-gray-950 gi-w-full gi-text-input" name="search_query"
+              <div class="gi-input-text-container  gi-flex-auto">
+                <div class="gi-input-text-container">
+                  <input placeholder="Enter search term" id="search" type="text"
+                    class="gi-border-gray-950 gi-w-full gi-input-text" name="search_query"
                     aria-label="Search the website">
                 </div>
               </div>
-              <button 
+              <button
                 class="gi-btn gi-btn-primary gi-btn-regular gi-ml-1 gi-flex-none">Search<span
                   aria-label="Search" role="img" style="
                   font-variation-settings:
