+++ apps/docs/content/4-patterns/2-pages/4-search-page.mdx
@@ -189,13 +189,13 @@ A template for a search page
         <form action="/search_page" class="gi-max-w-md gi-mx-auto">
           <h4 class="gi-heading-sm">Search the website</h4>
           <div class="gi-flex gi-items-end gi-mt-4">
-            <div class="gi-text-input-container gi-flex-auto">
-              <div class="gi-text-input-container">
+            <div class="gi-input-text-container gi-flex-auto">
+              <div class="gi-input-text-container">
                 <input
                   placeholder="Enter search term"
                   id="search"
                   type="text"
-                  class="gi-border-gray-950 gi-w-full gi-text-input"
+                  class="gi-border-gray-950 gi-w-full gi-input-text"
                   name="search_query"
                   aria-label="Search the website"
                 />
@@ -263,13 +263,13 @@ A template for a search page
             <form action="/search_page" class="gi-max-w-md gi-mx-auto">
               <h4 class="gi-heading-sm">Search the website</h4>
               <div class="gi-flex gi-items-end gi-mt-4">
-                <div class="gi-text-input-container gi-flex-auto">
-                  <div class="gi-text-input-container">
+                <div class="gi-input-text-container gi-flex-auto">
+                  <div class="gi-input-text-container">
                     <input
                       placeholder="Enter search term"
                       id="search"
                       type="text"
-                      class="gi-border-gray-950 gi-w-full gi-text-input"
+                      class="gi-border-gray-950 gi-w-full gi-input-text"
                       name="search_query"
                       aria-label="Search the website"
                     />
@@ -314,13 +314,13 @@ A template for a search page
                 <div class="grow gi-gap-4 gi-flex gi-flex-col">
                   <h2 class="gi-heading-lg">Search</h2>
 
-                  <div class="gi-text-input-container">
-                    <div class="gi-text-input-container {{ marginTop }}">
+                  <div class="gi-input-text-container">
+                    <div class="gi-input-text-container {{ marginTop }}">
                       <input
                         id="search-input"
                         type="text"
                         autocomplete="on"
-                        class="gi-border-gray-950 gi-w-full gi-text-input"
+                        class="gi-border-gray-950 gi-w-full gi-input-text"
                       />
                     </div>
                   </div>
@@ -557,17 +557,15 @@ A template for a search page
                 >
                   <div class="gi-combobox-search">
                     <div
-                      class="gi-text-input-container gi-combobox-search-input"
+                      class="gi-input-text-container gi-combobox-search-input"
                     >
-                      <div
-                        class="gi-text-input-container {{ marginTop }}"
-                      >
+                      <div class="gi-input-text-container {{ marginTop }}">
                         <input
                           id="input-id"
                           placeholder="Search"
                           type="text"
                           autocomplete="on"
-                          class="gi-border-gray-950 gi-w-full gi-text-input"
+                          class="gi-border-gray-950 gi-w-full gi-input-text"
                         />
                       </div>
                     </div>
@@ -889,17 +887,15 @@ A template for a search page
                 >
                   <div class="gi-combobox-search">
                     <div
-                      class="gi-text-input-container gi-combobox-search-input"
+                      class="gi-input-text-container gi-combobox-search-input"
                     >
-                      <div
-                        class="gi-text-input-container {{ marginTop }}"
-                      >
+                      <div class="gi-input-text-container {{ marginTop }}">
                         <input
                           id="input-id"
                           placeholder="Search"
                           type="text"
                           autocomplete="on"
-                          class="gi-border-gray-950 gi-w-full gi-text-input"
+                          class="gi-border-gray-950 gi-w-full gi-input-text"
                         />
                       </div>
                     </div>
@@ -1090,17 +1086,15 @@ A template for a search page
                 >
                   <div class="gi-combobox-search">
                     <div
-                      class="gi-text-input-container gi-combobox-search-input"
+                      class="gi-input-text-container gi-combobox-search-input"
                     >
-                      <div
-                        class="gi-text-input-container {{ marginTop }}"
-                      >
+                      <div class="gi-input-text-container {{ marginTop }}">
                         <input
                           id="input-id"
                           placeholder="Search"
                           type="text"
                           autocomplete="on"
-                          class="gi-border-gray-950 gi-w-full gi-text-input"
+                          class="gi-border-gray-950 gi-w-full gi-input-text"
                         />
                       </div>
                     </div>
