+++ apps/docs/content/4-patterns/2-pages/5-layout-page.mdx
@@ -224,15 +224,15 @@ A responsive grid that adjusts from 4 to 12 columns across screen sizes.
           <h4 class="gi-heading-sm">Search the website</h4>
 
           <div class="gi-flex gi-items-end gi-mt-4">
-            <div class="gi-text-input-container gi-flex-auto">
-              <div class="gi-text-input-container">
+            <div class="gi-input-text-container gi-flex-auto">
+              <div class="gi-input-text-container">
                 <input
                   id="search"
                   placeholder="Enter search term"
                   data-testid="textbox"
                   type="text"
                   autocomplete="on"
-                  class="gi-border-gray-950 gi-w-full gi-text-input"
+                  class="gi-border-gray-950 gi-w-full gi-input-text"
                 />
               </div>
             </div>
