+++ apps/docs/public/templates/basic-page.html
@@ -146,9 +146,9 @@
       <form action="/search_page" class="gi-max-w-md gi-mx-auto">
         <h4 class="gi-heading-sm">Search the website</h4>
         <div class="gi-flex gi-items-end gi-mt-4">
-          <div class="gi-text-input-container  gi-flex-auto">
-            <div class="gi-text-input-container"><input placeholder="Enter search term" id="search" type="text"
-                data-testid="textbox" class="gi-border-gray-950 gi-w-full gi-text-input" name="search_query"
+          <div class="gi-input-text-container  gi-flex-auto">
+            <div class="gi-input-text-container"><input placeholder="Enter search term" id="search" type="text"
+                data-testid="textbox" class="gi-border-gray-950 gi-w-full gi-input-text" name="search_query"
                 aria-label="Search the website"></div>
           </div><button data-testid="govieButton-undefined-undefined-undefined-"
             class="gi-btn gi-btn-primary gi-btn-regular gi-ml-1 gi-flex-none">Search<span data-testid="govie-icon"
@@ -194,10 +194,10 @@
           <form action="/search_page" class="gi-max-w-md gi-mx-auto">
             <h4 class="gi-heading-sm">Search the website</h4>
             <div class="gi-flex gi-items-end gi-mt-4">
-              <div class="gi-text-input-container  gi-flex-auto">
-                <div class="gi-text-input-container">
+              <div class="gi-input-text-container  gi-flex-auto">
+                <div class="gi-input-text-container">
                   <input placeholder="Enter search term" id="search" type="text" data-testid="textbox"
-                    class="gi-border-gray-950 gi-w-full gi-text-input" name="search_query"
+                    class="gi-border-gray-950 gi-w-full gi-input-text" name="search_query"
                     aria-label="Search the website">
                 </div>
               </div>
