+++ apps/docs/public/templates/search-page.html
@@ -201,14 +201,14 @@
         <form action="/search_page" class="gi-max-w-md gi-mx-auto">
           <h4 class="gi-heading-sm">Search the website</h4>
           <div class="gi-flex gi-items-end gi-mt-4">
-            <div class="gi-text-input-container gi-flex-auto">
-              <div class="gi-text-input-container">
+            <div class="gi-input-text-container gi-flex-auto">
+              <div class="gi-input-text-container">
                 <input
                   placeholder="Enter search term"
                   id="search"
                   type="text"
                   data-testid="textbox"
-                  class="gi-border-gray-950 gi-w-full gi-text-input"
+                  class="gi-border-gray-950 gi-w-full gi-input-text"
                   name="search_query"
                   aria-label="Search the website"
                 />
@@ -279,14 +279,14 @@
             <form action="/search_page" class="gi-max-w-md gi-mx-auto">
               <h4 class="gi-heading-sm">Search the website</h4>
               <div class="gi-flex gi-items-end gi-mt-4">
-                <div class="gi-text-input-container gi-flex-auto">
-                  <div class="gi-text-input-container">
+                <div class="gi-input-text-container gi-flex-auto">
+                  <div class="gi-input-text-container">
                     <input
                       placeholder="Enter search term"
                       id="search"
                       type="text"
                       data-testid="textbox"
-                      class="gi-border-gray-950 gi-w-full gi-text-input"
+                      class="gi-border-gray-950 gi-w-full gi-input-text"
                       name="search_query"
                       aria-label="Search the website"
                     />
@@ -333,14 +333,14 @@
                 <div class="grow gi-gap-4 gi-flex gi-flex-col">
                   <h2 class="gi-heading-lg">Search</h2>
 
-                  <div class="gi-text-input-container">
-                    <div class="gi-text-input-container {{ marginTop }}">
+                  <div class="gi-input-text-container">
+                    <div class="gi-input-text-container {{ marginTop }}">
                       <input
                         id="search-input"
                         data-testid="textbox"
                         type="text"
                         autocomplete="on"
-                        class="gi-border-gray-950 gi-w-full gi-text-input"
+                        class="gi-border-gray-950 gi-w-full gi-input-text"
                       />
                     </div>
                   </div>
@@ -585,16 +585,16 @@
                 >
                   <div class="gi-combobox-search">
                     <div
-                      class="gi-text-input-container gi-combobox-search-input"
+                      class="gi-input-text-container gi-combobox-search-input"
                     >
-                      <div class="gi-text-input-container {{ marginTop }}">
+                      <div class="gi-input-text-container {{ marginTop }}">
                         <input
                           id="input-id"
                           placeholder="Search"
                           data-testid="textbox"
                           type="text"
                           autocomplete="on"
-                          class="gi-border-gray-950 gi-w-full gi-text-input"
+                          class="gi-border-gray-950 gi-w-full gi-input-text"
                         />
                       </div>
                     </div>
@@ -923,16 +923,16 @@
                 >
                   <div class="gi-combobox-search">
                     <div
-                      class="gi-text-input-container gi-combobox-search-input"
+                      class="gi-input-text-container gi-combobox-search-input"
                     >
-                      <div class="gi-text-input-container {{ marginTop }}">
+                      <div class="gi-input-text-container {{ marginTop }}">
                         <input
                           id="input-id"
                           placeholder="Search"
                           data-testid="textbox"
                           type="text"
                           autocomplete="on"
-                          class="gi-border-gray-950 gi-w-full gi-text-input"
+                          class="gi-border-gray-950 gi-w-full gi-input-text"
                         />
                       </div>
                     </div>
@@ -1130,16 +1130,16 @@
                 >
                   <div class="gi-combobox-search">
                     <div
-                      class="gi-text-input-container gi-combobox-search-input"
+                      class="gi-input-text-container gi-combobox-search-input"
                     >
-                      <div class="gi-text-input-container {{ marginTop }}">
+                      <div class="gi-input-text-container {{ marginTop }}">
                         <input
                           id="input-id"
                           placeholder="Search"
                           data-testid="textbox"
                           type="text"
                           autocomplete="on"
-                          class="gi-border-gray-950 gi-w-full gi-text-input"
+                          class="gi-border-gray-950 gi-w-full gi-input-text"
                         />
                       </div>
                     </div>
