+++ apps/docs/src/components/document/common/mdx.tsx
@@ -65,6 +65,7 @@ import {
   Tag,
   TextArea,
   TextInput,
+  InputText,
   InputPassword,
   Tooltip,
   StepItem,
@@ -221,6 +222,7 @@ const documentComponents: MDXComponents = {
   Chip: (props) => <Chip {...props} />,
   IconGridLayout: (props) => <IconGridLayout {...props} />,
   TextInput: (props) => <TextInput {...props} />,
+  InputText,
   SectionBreak: (props) => <SectionBreak {...props} />,
   Select: (props) => <Select {...props} />,
   SelectItem: (props) => <SelectItem {...props} />,
