+++ examples/wagtail/core/jinja/vars.jinja
@@ -71,7 +71,7 @@
           "label": "Search",
           "keepOnMobile": true,
           "icon": "search",
-          "slot": "<form class='gi-max-w-md gi-mx-auto'><h4 class='gi-heading-sm'>Search the website</h4><div class='gi-flex gi-items-end gi-mt-4'><div class='gi-text-input-container gi-flex-auto'><div class='gi-text-input-container'><input placeholder='Enter search term' id='search' type='text' data-testid='textbox' class='gi-border-gray-950 gi-w-full gi-text-input' name='search_query' aria-label='Search the website'></div></div><button data-testid='govieButton-undefined-undefined-undefined-' class='gi-btn gi-btn-primary gi-btn-regular gi-ml-1 gi-flex-none'>Search<span data-testid='govie-icon' aria-label='Search' role='img' class='material-symbols-outlined gi-block' style='font-size: 24px; font-variation-settings: \"FILL\" 0, \"wght\" 400, \"GRAD\" 0, \"opsz\" 24;'>search</span></button></div></form>"
+          "slot": "<form class='gi-max-w-md gi-mx-auto'><h4 class='gi-heading-sm'>Search the website</h4><div class='gi-flex gi-items-end gi-mt-4'><div class='gi-input-text-container gi-flex-auto'><div class='gi-input-text-container'><input placeholder='Enter search term' id='search' type='text' data-testid='textbox' class='gi-border-gray-950 gi-w-full gi-input-text' name='search_query' aria-label='Search the website'></div></div><button data-testid='govieButton-undefined-undefined-undefined-' class='gi-btn gi-btn-primary gi-btn-regular gi-ml-1 gi-flex-none'>Search<span data-testid='govie-icon' aria-label='Search' role='img' class='material-symbols-outlined gi-block' style='font-size: 24px; font-variation-settings: \"FILL\" 0, \"wght\" 400, \"GRAD\" 0, \"opsz\" 24;'>search</span></button></div></form>"
         },
         {
           "href": "#",
