+++ packages/design/tailwind/css/components.css
@@ -31,6 +31,6 @@
 @import './components/tabs.css';
 @import './components/tag.css';
 @import './components/text-area.css';
-@import './components/text-input.css';
+@import './components/input-text.css';
 @import './components/toast.css';
 @import './components/tootlip.css';
