+++ packages/design/tailwind/css/components/input-text.css
@@ -1,62 +1,62 @@
-.gi-text-input-inner {
+.gi-input-text-inner {
   @apply gi-w-full gi-flex gi-items-center gi-relative;
 }
 
-.gi-text-input-container {
+.gi-input-text-container {
   @apply gi-w-full gi-flex gi-items-center gi-relative;
 }
 
-.gi-text-input-prefix {
+.gi-input-text-prefix {
   @apply gi-rounded-l-sm xs:gi-text-md gi-text-sm gi-grid gi-place-content-center gi-bg-gray-100 gi-inline-block gi-flex-[0_0_auto] gi-text-center gi-whitespace-nowrap gi-cursor-default gi-px-2 gi-py-2 gi-border-l-sm gi-border-t-sm gi-border-b-sm gi-border-solid gi-border-gray-950 gi-min-w-10 gi-h-10;
 }
-.gi-text-input-prefix[data-disabled='true'] {
+.gi-input-text-prefix[data-disabled='true'] {
   @apply gi-border-gray-400 gi-text-gray-700 gi-cursor-not-allowed;
 }
 
-.gi-text-input-suffix {
+.gi-input-text-suffix {
   @apply gi-rounded-r-sm xs:gi-text-md gi-text-sm gi-grid gi-place-content-center gi-bg-gray-100 gi-inline-block gi-flex-[0_0_auto] gi-text-center gi-whitespace-nowrap gi-cursor-default gi-px-2 gi-py-2 gi-border-r-sm gi-border-t-sm gi-border-b-sm gi-border-solid gi-border-gray-950 gi-min-w-10 gi-h-10;
 }
-.gi-text-input-suffix[data-disabled='true'] {
+.gi-input-text-suffix[data-disabled='true'] {
   @apply gi-border-gray-400 gi-text-gray-700 gi-cursor-not-allowed;
 }
 
-.gi-text-input {
+.gi-input-text {
   @apply gi-w-full gi-rounded-sm gi-focus-state-outline gi-focus-state-border gi-flex-initial gi-border-sm gi-border-solid gi-box-border gi-h-10 gi-mt-0 gi-p-2 gi-z-1 xs:gi-text-md gi-text-sm gi-leading-10 xs:gi-leading-5 gi-input-disabled-state gi-border-gray-950;
 }
-.gi-text-input[data-icon-start='true'][data-prefix='true'] {
+.gi-input-text[data-icon-start='true'][data-prefix='true'] {
   @apply gi-pl-9;
 }
 
-.gi-text-input[data-icon-start='true'] {
+.gi-input-text[data-icon-start='true'] {
   @apply gi-pl-9;
 }
 
-.gi-text-input[data-icon-end='false'][data-end-element='true'] {
+.gi-input-text[data-icon-end='false'][data-end-element='true'] {
   @apply gi-pr-10;
 }
-.gi-text-input[data-icon-end='true'][data-end-element='true'] {
+.gi-input-text[data-icon-end='true'][data-end-element='true'] {
   @apply gi-pr-18;
 }
-.gi-text-input[data-icon-end='true'][data-end-element='false'] {
+.gi-input-text[data-icon-end='true'][data-end-element='false'] {
   @apply gi-pr-8;
 }
 
-.gi-text-input-icon-end {
+.gi-input-text-icon-end {
   @apply gi-absolute gi-z-200 gi-right-2;
 }
 
-.gi-text-input-icon-end[data-end-element='true'] {
+.gi-input-text-icon-end[data-end-element='true'] {
   @apply gi-right-11;
 }
-.gi-text-input-icon-end[data-end-element='true'][data-suffix='true'] {
+.gi-input-text-icon-end[data-end-element='true'][data-suffix='true'] {
   @apply gi-right-10;
 }
 
-.gi-text-input-icon-start {
+.gi-input-text-icon-start {
   @apply gi-absolute gi-z-200 gi-left-2;
 }
 
-.gi-text-input-end-element {
+.gi-input-text-end-element {
   @apply gi-flex gi-absolute gi-z-200 gi-right-1;
 }
 .gi-input-half-width {
