+++ packages/html/ds/src/text-input/text-input.stories.ts
@@ -121,51 +121,51 @@ export const ResponsiveLayout: Story = {
   <div class="md:gi-w-2/3 gi-w-full">
    <div class="gi-flex gi-w-full gi-justify-start gi-items-start gi-flex-col gi-gap-3 gi-flex-nowrap" role="region" aria-label="Items Stacked" data-testid="govie-stack" style="height: 100%;">
       <div class="gi-flex gi-w-full gi-justify-start gi-items-start gi-flex-col md:gi-flex-row gi-gap-3 gi-flex-nowrap" role="region" aria-label="Items Stacked" data-testid="govie-stack" style="height: 100%;">
-         <div role="group" class="gi-text-input-container" aria-labelledby="text-1-label">
+         <div role="group" class="gi-input-text-container" aria-labelledby="text-1-label">
             <label class="gi-text-md gi-label gi-mb-1" for="text-1" aria-label="First Name" id="text-1-label">First Name</label>
             <div class="gi-hint-text-md gi-hint-text " aria-label="Your first name." id="text-1-hint">Your first name.</div>
-            <div class="gi-text-input-container"><input id="text-1" type="text" data-testid="govie-stack-item-0" aria-invalid="false" class="gi-text-input gi-border-gray-950"></div>
+            <div class="gi-input-text-container"><input id="text-1" type="text" data-testid="govie-stack-item-0" aria-invalid="false" class="gi-input-text gi-border-gray-950"></div>
          </div>
-         <div role="group" class="gi-text-input-container" aria-labelledby="text-2-label">
+         <div role="group" class="gi-input-text-container" aria-labelledby="text-2-label">
             <label class="gi-text-md gi-label gi-mb-1" for="text-2" aria-label="Last Name" id="text-2-label">Last Name</label>
             <div class="gi-hint-text-md gi-hint-text " aria-label="Your last name." id="text-2-hint">Your last name.</div>
-            <div class="gi-text-input-container"><input id="text-2" type="text" data-testid="govie-stack-item-1" aria-invalid="false" class="gi-text-input gi-border-gray-950"></div>
+            <div class="gi-input-text-container"><input id="text-2" type="text" data-testid="govie-stack-item-1" aria-invalid="false" class="gi-input-text gi-border-gray-950"></div>
          </div>
       </div>
       <div class="gi-flex gi-w-full gi-justify-start gi-items-start gi-flex-col md:gi-flex-row gi-gap-3 gi-flex-nowrap" role="region" aria-label="Items Stacked" data-testid="govie-stack" style="height: 100%;">
-         <div role="group" class="gi-text-input-container" aria-labelledby="text-4-label">
+         <div role="group" class="gi-input-text-container" aria-labelledby="text-4-label">
             <label class="gi-text-md gi-label gi-mb-1" for="text-4" aria-label="Address" id="text-4-label">Address</label>
             <div class="gi-hint-text-md gi-hint-text " aria-label="Where you live." id="text-4-hint">Where you live.</div>
-            <div class="gi-text-input-container"><input id="text-4" type="text" data-testid="govie-stack-item-0" aria-invalid="false" class="gi-text-input gi-border-gray-950" maxlength="5"></div>
+            <div class="gi-input-text-container"><input id="text-4" type="text" data-testid="govie-stack-item-0" aria-invalid="false" class="gi-input-text gi-border-gray-950" maxlength="5"></div>
          </div>
       </div>
       <div class="gi-flex gi-w-full gi-justify-start gi-items-start gi-flex-col md:gi-flex-row gi-gap-3 gi-flex-nowrap" role="region" aria-label="Items Stacked" data-testid="govie-stack" style="height: 100%;">
-         <div role="group" class="gi-text-input-container" aria-labelledby="text-input-id-label">
+         <div role="group" class="gi-input-text-container" aria-labelledby="text-input-id-label">
             <label class="gi-text-md gi-label gi-mb-1" for="text-input-id" aria-label="Date of birth" id="text-input-id-label">Date of birth</label>
             <div class="gi-hint-text-md gi-hint-text " aria-label="Your date of birth." id="text-input-id-hint">Your date of birth.</div>
-            <div class="gi-text-input-container"><input id="text-input-id" type="date" data-testid="govie-stack-item-0" aria-invalid="false" class="gi-text-input gi-border-gray-950"></div>
+            <div class="gi-input-text-container"><input id="text-input-id" type="date" data-testid="govie-stack-item-0" aria-invalid="false" class="gi-input-text gi-border-gray-950"></div>
          </div>
-         <div role="group" class="gi-text-input-container" aria-labelledby="text-input-id-label">
+         <div role="group" class="gi-input-text-container" aria-labelledby="text-input-id-label">
             <label class="gi-text-md gi-label gi-mb-1" for="text-input-id" aria-label="Height" id="text-input-id-label">Height</label>
             <div class="gi-hint-text-md gi-hint-text " aria-label="Your height" id="text-input-id-hint">Your height</div>
-            <div class="gi-text-input-container">
-               <div class="gi-text-input-prefix">cm</div>
-               <input id="text-input-id" type="text" data-testid="govie-stack-item-1" aria-invalid="false" class="gi-text-input gi-border-gray-950">
+            <div class="gi-input-text-container">
+               <div class="gi-input-text-prefix">cm</div>
+               <input id="text-input-id" type="text" data-testid="govie-stack-item-1" aria-invalid="false" class="gi-input-text gi-border-gray-950">
             </div>
          </div>
          <div class="gi-w-full sm:gi-w-[80px] gi-flex-none" data-testid="govie-stack-item-2">
-            <div role="group" class="gi-text-input-container" aria-labelledby="text-4-label">
+            <div role="group" class="gi-input-text-container" aria-labelledby="text-4-label">
                <label class="gi-text-md gi-label gi-mb-1" for="text-4" aria-label="Age" id="text-4-label">Age</label>
                <div class="gi-hint-text-md gi-hint-text " aria-label="Your Age." id="text-4-hint">Your Age.</div>
-               <div class="gi-text-input-container"><input id="text-4" type="text" data-testid="textbox" aria-invalid="false" class="gi-text-input gi-border-gray-950" maxlength="3"></div>
+               <div class="gi-input-text-container"><input id="text-4" type="text" data-testid="textbox" aria-invalid="false" class="gi-input-text gi-border-gray-950" maxlength="3"></div>
             </div>
          </div>
       </div>
-      <div role="group" class="gi-text-input-container gi-error-state" aria-labelledby="text-4-label">
+      <div role="group" class="gi-input-text-container gi-error-state" aria-labelledby="text-4-label">
          <label class="gi-text-md gi-label gi-mb-1" for="text-4" aria-label="Phone Number" id="text-4-label">Phone Number</label>
          <div class="gi-hint-text-md gi-hint-text " aria-label="Your phone number." id="text-4-hint">Your phone number.</div>
          <div role="alert" class="gi-error-text-md gi-error-text " id="text-4-error">Error: Please correct this issue.</div>
-         <div class="gi-text-input-container"><input id="text-4" type="text" data-testid="govie-stack-item-3" aria-invalid="true" class="gi-text-input gi-border-red-600" pattern="d*" maxlength="10"></div>
+         <div class="gi-input-text-container"><input id="text-4" type="text" data-testid="govie-stack-item-3" aria-invalid="true" class="gi-input-text gi-border-red-600" pattern="d*" maxlength="10"></div>
       </div>
    </div>
 </div>
@@ -277,38 +277,38 @@ export const WithHalfWidth: Story = {
 // export const AllStates: Story = {
 //   render: () => `
 //   <div class="gi-gap-4">
-//   <div class="gi-text-input-container">
+//   <div class="gi-input-text-container">
 //     <label class="gi-text-md gi-label gi-mb-2" for="default-input">Default</label>
-//     <div class="gi-text-input-container">
+//     <div class="gi-input-text-container">
 //       <input
 //         id="default-input"
 //         type="text"
 //         data-testid="textbox"
-//         class="gi-border-gray-950 gi-w-full gi-text-input"
+//         class="gi-border-gray-950 gi-w-full gi-input-text"
 //       />
 //     </div>
 //   </div>
 
-//   <div class="gi-text-input-container">
+//   <div class="gi-input-text-container">
 //     <label class="gi-text-md gi-label gi-mb-2" for="focus-input">Focus</label>
-//     <div class="gi-text-input-container">
+//     <div class="gi-input-text-container">
 //       <input
 //         id="focus-input"
 //         type="text"
 //         data-testid="textbox"
-//         class="gi-border-gray-950 gi-w-full gi-text-input pseudo-focus"
+//         class="gi-border-gray-950 gi-w-full gi-input-text pseudo-focus"
 //       />
 //     </div>
 //   </div>
 
-//   <div class="gi-text-input-container">
+//   <div class="gi-input-text-container">
 //     <label class="gi-text-md gi-label gi-mb-2" for="input-disabled">Disabled</label>
-//     <div class="gi-text-input-container">
+//     <div class="gi-input-text-container">
 //       <input
 //         id="input-disabled"
 //         type="text"
 //         data-testid="textbox"
-//         class="gi-border-gray-950 gi-w-full gi-text-input gi-text-input-disabled"
+//         class="gi-border-gray-950 gi-w-full gi-input-text gi-input-text-disabled"
 //         disabled
 //       />
 //     </div>
