+++ packages/react/ds/src/index.ts
@@ -117,6 +117,8 @@ export { TabPanel } from './tabs/tab-panel.js';
 export { Tabs, type TabsProps } from './tabs/tabs.js';
 export { Tag, type TagProps, TagTypeEnum, type TagType } from './tag/tag.js';
 export { TextInput, type TextInputProps } from './text-input/text-input.js';
+export { InputText } from './input-text/input-text.js';
+export type { InputProps, InputActionButtonProps } from './input-text/type.js';
 export {
   InputPassword,
   type InputPasswordProps,
