+++ packages/react/ds/src/input-text/type.ts
@@ -1,0 +1,34 @@
+import React from 'react';
+import { IconId } from '../icon/icon.js';
+
+export type InputActionButtonProps = {
+  icon: IconId;
+  onClick: (event: React.MouseEvent<HTMLButtonElement>) => void;
+  disabled?: boolean;
+  ariaLabel?: string;
+  dataTestId?: string;
+  ref?: React.Ref<HTMLButtonElement>;
+};
+
+export type InputProps = React.InputHTMLAttributes<HTMLInputElement> & {
+  prefix?: React.ReactNode;
+  suffix?: React.ReactNode;
+  inputClassName?: string;
+  iconStart?: IconId;
+  iconEnd?: IconId;
+  inputActionButton?: InputActionButtonProps;
+  type?:
+    | 'text'
+    | 'date'
+    | 'datetime-local'
+    | 'email'
+    | 'month'
+    | 'number'
+    | 'password'
+    | 'tel'
+    | 'time'
+    | 'url'
+    | 'week';
+  halfFluid?: boolean;
+  clearButtonEnabled?: boolean;
+};
